requirements.txt
gitdb2==2.0.5
GitPython==2.1.13
nltk==3.4.4
pkg-resources==0.0.0
pyspellchecker==0.5.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
termcolor==1.1.0
